requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.3.3
langchain-openai==0.2.2
langgraph==0.2.27
httpx==0.27.2
moviepy==1.0.3
numpy==1.26.4
scikit-learn==1.5.2
pydub==0.25.1
tqdm==4.66.5
boto3==1.35.26
google-cloud-storage==2.18.2
loguru==0.7.2
jinja2==3.1.4
